comment create,read,delete

+++ src/main/java/spring/board/controller/boardcomment/BoardCommentController.java
@@ -2,11 +2,14 @@ package spring.board.controller.boardcomment;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.util.StringUtils;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import spring.board.domain.Member;
+import spring.board.dto.boardcomment.PostAuthRequest;
 import spring.board.dto.boardcomment.SaveCommentDto;
 import spring.board.service.BoardCommentService;
 
@@ -35,11 +38,29 @@ public class BoardCommentController {
     }
 
     @GetMapping("/comment/{commentId}/delete") // 댓글의 PK값으로 댓글을 찾아서 삭제
-    public String deleteComment(@PathVariable Long commentId){
+    public String deleteComment(@PathVariable Long commentId, Model model){
 
-         Long boardId = boardCommentService.deleteComment(commentId);
-         // 삭제하고 그 게시물의 번호를 리턴
+        PostAuthRequest userInfoForComment = boardCommentService.getUserInfoForComment(commentId);
+        String username = userInfoForComment.getUsername();
 
-        return "redirect:/contents/" + boardId;
+        log.info("commentUsername={}",userInfoForComment.getUsername());
+        log.info("authUsername={}",getAuthMember().getUsername());
+
+        if(!username.equals(getAuthMember().getUsername())){
+            model.addAttribute("message","다른 사용자의 댓글입니다.");
+            model.addAttribute("searchUrl","/contents/" + userInfoForComment.getBoardId());
+            return "warningPage";
+        }
+
+        // 삭제하고 그 게시물의 번호를 리턴
+        boardCommentService.deleteComment(commentId);
+
+        return "redirect:/contents/" + userInfoForComment.getBoardId();
+    }
+
+    private static Member getAuthMember() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Member member = (Member) authentication.getPrincipal();
+        return member;
     }
 }
